feat: seperate out config into its own crate
