+++ crates/config/Cargo.toml
@@ -1,0 +1,55 @@
+[package]
+name = "prism-config"
+name.workspace = true
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+description.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+keywords.workspace = true
+readme.workspace = true
+
+[dependencies]
+axum = { workspace = true }
+borsh = { workspace = true }
+tower-http = { workspace = true }
+utoipa = { workspace = true }
+utoipa-swagger-ui = { workspace = true }
+async-trait = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+redis = { workspace = true }
+ed25519-dalek = { workspace = true }
+ed25519 = { workspace = true }
+base64 = { workspace = true }
+tokio = { workspace = true }
+bellman = { workspace = true }
+bincode = { workspace = true }
+bls12_381 = { workspace = true }
+rand = { workspace = true }
+hex = { workspace = true }
+ff = { workspace = true }
+log = { workspace = true }
+pretty_env_logger = { workspace = true }
+clap = { workspace = true }
+config = { workspace = true }
+thiserror = { workspace = true }
+indexed-merkle-tree = { workspace = true }
+dotenvy = { workspace = true }
+celestia-rpc = { workspace = true }
+celestia-types = { workspace = true }
+mockall = { workspace = true }
+keystore-rs = { workspace = true }
+toml = { workspace = true }
+dirs = { workspace = true }
+anyhow = { workspace = true }
+jmt = { workspace = true }
+sha2 = { workspace = true }
+auto_impl = { workspace = true }
+prism-common = { workspace = true }
+prism-storage = { workspace = true }
+prism-errors = { workspace = true }
+prism-groth16 = { workspace = true }
+sp1-sdk = { workspace = true }
