+++ crates/prism/src/main.rs
@@ -19,6 +19,8 @@ extern crate log;
 
 pub const PRISM_ELF: &[u8] = include_bytes!("../../../elf/riscv32im-succinct-zkvm-elf");
 
+
+
 /// The main function that initializes and runs a prism client.
 #[tokio::main()]
 async fn main() -> std::io::Result<()> {
